rename setting.py to esp32_settings.py
