+++ examples/esp32spi_aio_post.py
@@ -12,7 +12,7 @@ print("ESP32 SPI webclient test")
 
 # Get wifi details and more from a settings.py file
 try:
-    from settings import settings
+    from esp32spi_settings import settings
 except ImportError:
     print("WiFi settings are kept in settings.py, please add them there!")
     raise
