+++ examples/esp32spi_settings.py
